+++ serialization/src/commonMain/kotlin/nl/adaptivity/xmlutil/serialization/structure/XmlOrderConstraint.kt
@@ -29,4 +29,8 @@ package nl.adaptivity.xmlutil.serialization.structure
  */
 public data class XmlOrderConstraint(val before: Int, val after: Int) {
     public inline fun <R> map(transform: (Int) -> R): Pair<R, R> = Pair(transform(before), transform(after))
+
+    public companion object {
+        public const val OTHERS: Int = -2;
+    }
 }