+++ Changelog.md
@@ -1,3 +1,21 @@
+# 0.86.2-SNAPSHOT
+Features:
+- Support wildcard (`*`) for `XmlBefore` and `XmlAfter`. These create a
+  partition in ordering. An element/attribute with wildcard before will
+  be ordered before/after elements that do not have this. Explicit order
+  relationships will be maintained.
+- Support checking ordering in reading.
+- Support ID type attributes with an @XmlId attribute marking them
+- Add support for strict attribute name matching (not allowing null namespace)
+  matchup
+
+Fixes:
+- For attribute lists, make sure to collapse the whitespace.
+- When attributes have an `@XmlSerialName` annotation with a default namespace
+  value, then this will result in a non-qualified attribute.
+- Fix nullable QName serialization (and probably other nullable inline-like)
+  serialization.
+
 # 0.86.1
 *(July 5, 2023)<br />*
 Features: